add: partial task management ui and state
Adds the task manager UI for Archive and Unarchive a task, including the buttons required to navigate between archive list and the task manager. Also includes the features of a task removal from a column and task reappearance when archived and unarchived, respectively. This commit fulfils the RQ6.8 and RQ6.9 of the SRS document.