add: notes side bar CCA and Society views
Creates CCA and Society views conditionally for the NotesSideBar for the Form Viewer, so
that the CCA can update their note and the society can edit theirs depending on the isCCA boolean.
It does not update to the backend, add changes in state only and relies on isCCA instead of proper
verification of role. A 300 characters restriction included on the CCANotesSideBar as well.